// src/ipc/generated/tasks/Applied.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { BoardRev } from "./BoardRev";

/**
 * How an applied change went.
 */
export type Applied = { "kind": "applied", revs: Array<BoardRev>, } | { "kind": "alreadyApplied", revs: Array<BoardRev>, };